// src/lib/types.ts
export type UserRole = 
  | "district_engineer" 
  | "district_manager"
  | "regional_engineer" 
  | "regional_general_manager"
  | "global_engineer" 
  | "technician" 
  | "system_admin"
  | "ict"
  | "load_monitoring_edit"
  | "load_monitoring_delete"
  | "admin"
  | "project_engineer"
  | "ashsubt"
  | "accsubt"
  | "pending"
  | null;

export interface User {
  id: string;
  uid: string;
  email: string;
  displayName: string;
  name: string;
  role: UserRole;
  region?: string;
  regionId?: string;
  district?: string;
  districtId?: string;
  tempPassword?: string;
  mustChangePassword?: boolean;
  password?: string;
  staffId?: string;
  disabled?: boolean;
  photoURL?: string;
  status?: "pre_registered" | "active" | "inactive";
  createdAt?: any;
  updatedAt?: any;
}